requirements.txt
google-generativeai>=0.3.0
google-adk>=0.1.0
python-dotenv>=1.0.0


